prompt_project/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
google-generativeai==0.3.2
python-jose==3.3.0
Werkzeug==3.0.1 